Chores change route name and test review routes
