+++ database/dev-initdb.b/001-init-dev.sql
@@ -68,3 +68,30 @@ CREATE TABLE IF NOT EXISTS shop_kind(
     FOREIGN KEY (id_shop) REFERENCES shops(id_shop) ON DELETE CASCADE ON UPDATE CASCADE,
     FOREIGN KEY (id_kind) REFERENCES kinds(id_kind) ON DELETE CASCADE ON UPDATE CASCADE
 );
+
+-- Table: Horaires d'ouverture
+CREATE TABLE IF NOT EXISTS openings (
+    id INT PRIMARY KEY NOT NULL AUTO_INCREMENT,
+    id_day INTEGER NOT NULL,
+    id_shop INTEGER NOT NULL,
+    open VARCHAR(255) NOT NULL,
+    close VARCHAR(255) NOT NULL,
+    created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    FOREIGN KEY (id_shop) REFERENCES shops(id_shop) on delete cascade on update cascade
+);
+
+
+CREATE TABLE IF NOT EXISTS benefits (
+    id_benefit INTEGER NOT NULL AUTO_INCREMENT,
+    id_shop INTEGER NOT NULL,
+    name VARCHAR(255) NOT NULL,
+    description VARCHAR(255) NOT NULL,
+    duration VARCHAR(255) NOT NULL,
+    price VARCHAR(255) NOT NULL,
+    created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    PRIMARY KEY (id_benefit),
+    FOREIGN KEY (id_shop) REFERENCES shops(id_shop) ON DELETE CASCADE ON UPDATE CASCADE
+);
+
