+++ go_app/structure/collaborator_structure.go
