+++ go_app/structure/openingHours_structure.go
